Разделяет один JS-файл на несколько модулей
