+++ js/const/const.js
@@ -1,13 +1,13 @@
-export const COUNT_PUBLISH_PHOTO = 25;
-export const MAX_COUNT_ID = 25;
-export const COMMENTS_MAX_COUNT = 5;
-export const LIKES_MIN_COUNT = 15;
-export const LIKES_MAX_COUNT = 200;
-export const AVATARS_MAX_COUNT = 6;
+const COUNT_PUBLISH_PHOTO = 25;
+const MAX_COUNT_ID = 25;
+const COMMENTS_MAX_COUNT = 5;
+const LIKES_MIN_COUNT = 15;
+const LIKES_MAX_COUNT = 200;
+const AVATARS_MAX_COUNT = 6;
 
 
 //массив строк описаний фотографий
-export const DESCRIPTIONS = [
+const DESCRIPTIONS = [
   'Наслаждайтесь каждым моментом.',
   'Назад пути нет!',
   'Сколько мне ещё предстоит рассказать...',
@@ -15,11 +15,11 @@ export const DESCRIPTIONS = [
   'Повод задуматься.',
 ];
 
-export const MESSAGES = [
+const MESSAGES = [
   'В целом всё неплохо. Но не всё.',
   'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',];
 
-export const NAMES = [
+const NAMES = [
   'Екатерина',
   'Ксения',
   'Арсений',
@@ -27,3 +27,5 @@ export const NAMES = [
   'Александр',
   'Марина',
 ];
+
+export {COUNT_PUBLISH_PHOTO, MAX_COUNT_ID, COMMENTS_MAX_COUNT, LIKES_MIN_COUNT, LIKES_MAX_COUNT, AVATARS_MAX_COUNT, DESCRIPTIONS, MESSAGES, NAMES};
