+++ js/utils/helpers.js
@@ -1,5 +1,5 @@
 // функция нахождения рандомного числа из заданного диапазона
-export const getRandomInteger = (a, b) => {
+const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
   const result = Math.random() * (upper - lower + 1) + lower;
@@ -10,7 +10,7 @@ export const getRandomInteger = (a, b) => {
 const arrayIds = [];
 
 // функция нахождения случайного уникального числа
-export function getUniqueId (min, max) {
+function getUniqueId (min, max) {
   let uniqueId = getRandomInteger(min, max);
 
   while ((arrayIds.length < (max - min + 1)) || (arrayIds.includes(uniqueId) === false)) {
@@ -23,4 +23,6 @@ export function getUniqueId (min, max) {
 }
 
 // функция нахождения рандомного элемента массива
-export const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, getUniqueId, getRandomArrayElement};
